Fix lifeforms coalescing in the top-left corner
Root cause: recomputeCenter zeroed a dead cell''s center to [0,0] - the
top-left corner - and two systems consumed that corpse coordinate:

- The splitter on-death handler read the already-zeroed center, so every
  dead splitter spawned its two offspring AT the corner, repeatedly
  injecting biomass there.
- chooseEcosystemTarget never checked the control sample was alive, so
  once it died at [0,0] any lifeform within threat range locked onto the
  corpse forever. Predators chasing corner prey cascaded the whole dish
  into the corner.

Dead cells now retain their last known center, and targeting ignores a
dead control sample. Regression tests cover both paths.

Also widen the grid to Uint16Array: cell ids grow without reuse during
an epoch, and a Uint8Array wraps id 256 to empty and 257 to the control
sample, silently corrupting pixel ownership in long sessions.

+++ src/sim/grid.ts
@@ -4,7 +4,10 @@ export function createGrid(LX: number, LY: number, wrap: boolean): Grid {
   return {
     LX,
     LY,
-    cells: new Uint8Array(LX * LY),
+    // Uint16: cell ids grow without reuse during an epoch (births, reseeds,
+    // outbreaks). A Uint8Array would wrap id 256 to 0 (empty) and 257 to 1
+    // (the control sample), silently corrupting pixel ownership.
+    cells: new Uint16Array(LX * LY),
     boundary: new Set<number>(),
     wrap,
   };
